Restructure augment code
