+++ common/src/main/java/rearth/oritech/OritechClient.java
@@ -10,7 +10,7 @@ import net.minecraft.client.option.KeyBinding;
 import net.minecraft.client.util.InputUtil;
 import net.minecraft.entity.EquipmentSlot;
 import org.lwjgl.glfw.GLFW;
-import rearth.oritech.block.entity.interaction.PlayerModifierTestEntity;
+import rearth.oritech.block.entity.augmenter.PlayerAugments;
 import rearth.oritech.client.init.ModRenderers;
 import rearth.oritech.client.init.ModScreens;
 import rearth.oritech.client.other.OreFinderRenderer;
@@ -57,7 +57,7 @@ public final class OritechClient {
         
         ClientTickEvents.START_CLIENT_TICK.register(client -> {
             if (client.player != null)
-                PlayerModifierTestEntity.clientTickAugments(client.player);
+                PlayerAugments.clientTickAugments(client.player);
         });
     }
