+++ common/src/main/java/rearth/oritech/client/ui/PlayerModifierScreenHandler.java
@@ -9,7 +9,7 @@ import net.minecraft.item.ItemStack;
 import net.minecraft.screen.ScreenHandler;
 import net.minecraft.util.math.BlockPos;
 import org.jetbrains.annotations.NotNull;
-import rearth.oritech.block.entity.interaction.PlayerModifierTestEntity;
+import rearth.oritech.block.entity.augmenter.PlayerModifierTestEntity;
 import rearth.oritech.client.init.ModScreens;
 
 public class PlayerModifierScreenHandler extends ScreenHandler {
